Major syntax change: nJ(requirements)(obj)
Create a schema by passing the requirements object to nJ, which is
obtained by require('normalize-json')

The schema is a function which accepts a js object. It does not
modify the original object in normalizing, but rather returns the
modified, normalized object. Also, rather than returning error
objects, it throws errors.

Refactored jasmine matchers into module.exports so they can be
used in other projects.